Split libc and linux_raw implementations for program into separate modules
